Validate some methods in Ticket service

+++ VDG_Backend/VDG_Web_Api/src/Services/TicketService.cs
@@ -3,23 +3,40 @@ using VDG_Web_Api.src.Models;
 using VDG_Web_Api.src.Repositories.Interfaces;
 using VDG_Web_Api.src.Services.Interfaces;
 
+
 namespace VDG_Web_Api.src.Services
 {
     public class TicketService : ITicketService
     {
+        private readonly int UpdateAndDeleteThreshold = 5;
         private readonly ITicketRepository _ticketRepository;
         private readonly IUserService _userService;
         public readonly IDoctorService _doctorService;
+        private readonly IUserRepository _userRepository;
 
 
-        public TicketService(ITicketRepository ticketRepository, IUserService userService, IDoctorService doctorService)
+        public TicketService(ITicketRepository ticketRepository, IUserService userService, IDoctorService doctorService, IUserRepository userRepository)
         {
             this._ticketRepository = ticketRepository;
             this._userService = userService;
             this._doctorService = doctorService;
+            this._userRepository = userRepository;
         }
+        // Done 
         public async Task DeleteMessageAsync(int id)
         {
+            TicketMessage ticketMessage = await _ticketRepository.GetTicketMessageAsync(id);
+
+            if (ticketMessage == null)
+            {
+                throw new ArgumentNullException($"No such ticket with this id: {id}");
+            }
+
+            if (DateTime.Now.Subtract(ticketMessage.Date).Minutes > UpdateAndDeleteThreshold)
+            {
+                throw new Exception($"The minimum time to remove the message has passed");
+            }
+
             try
             {
                 await _ticketRepository.DeleteMessageAsync(id);
@@ -29,25 +46,54 @@ namespace VDG_Web_Api.src.Services
                 throw new InvalidOperationException($"Faild to delete the ticket meassge, Error: {ex.Message}", ex);
             }
         }
+        // Done
+        public async Task UpdateMessageAsync(TicketMessageDTO ticketMessageDTO)
+        {
+
+            var ticketMessage = MapToTicketMessage(ticketMessageDTO);
+            if (ticketMessage == null)
+            {
+                throw new ArgumentNullException($"No such Message");
+            }
+            if (DateTime.Now.Subtract(ticketMessage.Date).Minutes > UpdateAndDeleteThreshold)
+            {
+                throw new Exception($"The minimum time to remove the message has passed");
+            }
+
+            try
+            {
+                await _ticketRepository.UpdateMessageAsync(ticketMessage);
+            }
+            catch (Exception ex)
+            {
+
+                throw new InvalidOperationException($"Could not update, Error: {ex.Message}", ex);
+            }
+        }
 
         public async Task<IEnumerable<DoctorTicketDTO>> GetDoctorConsultationsAsync(string doctorId)
         {
+
+            if (doctorId == null)
+            {
+                throw new ArgumentNullException($"No such doctor with this id: {doctorId}");
+            }
             try
             {
                 var doctorConsultations = await _ticketRepository.GetConsultationsAsync(doctorId, null);
 
-                var userIds = doctorConsultations.Select(r => r.UserId)
-                .Distinct()
-                .ToList();
-
-                var userDtos = await Task.WhenAll(userIds.Where(Id => Id != null).Select(Id => _userService.GetUser(Id!.Value)));
                 return doctorConsultations.Select(d =>
                 {
                     var ticketDto = MapToTicketDto(d);
 
-                    var userDto = userDtos.FirstOrDefault(user => user?.Id == ticketDto.UserId);
+                    var user =
+                    /* Todo 
+                     
+                        get Fname and Lname 
 
-                    return new DoctorTicketDTO() { TicketDto = ticketDto, UserDto = userDto };
+                    */
+
+                    return new DoctorTicketDTO() { TicketDto = ticketDto, Fname = "", Lname = "" };
                 }).ToList();
 
 
@@ -104,7 +150,6 @@ namespace VDG_Web_Api.src.Services
             }
         }
 
-        // N 
         public async Task SendMessageAsync(TicketMessageDTO ticketMessageDto)
         {
             try
@@ -118,20 +163,6 @@ namespace VDG_Web_Api.src.Services
                 throw new InvalidOperationException($"Could not send, Error: {ex.Message}", ex);
             }
         }
-        public async Task UpdateMessageAsync(TicketMessageDTO ticketMessageDTO)
-        {
-            try
-            {
-                var ticketMessage = MapToTicketMessage(ticketMessageDTO);
-                await _ticketRepository.UpdateMessageAsync(ticketMessage);
-            }
-            catch (Exception ex)
-            {
-
-                throw new InvalidOperationException($"Could not update, Error: {ex.Message}", ex);
-            }
-        }
-
 
         // Mapping 
 
@@ -139,11 +170,12 @@ namespace VDG_Web_Api.src.Services
         {
             return new TicketDTO()
             {
+                CloseDate = ticket.CloseDate,
                 Id = ticket.Id,
                 DoctorId = ticket.DoctorId,
-                UserId = ticket.UserId,
                 Status = ticket.Status,
-                CloseDate = ticket.CloseDate,
+                UserId = ticket.UserId,
+
             };
         }
 
@@ -171,5 +203,9 @@ namespace VDG_Web_Api.src.Services
             };
         }
 
+        public Task DeleteMessageAsync(int id, DateTime date)
+        {
+            throw new NotImplementedException();
+        }
     }
 }
